added home buttons and censored password
Added a button to show/hide the assignments and another button that both shows and reads the assignments. They are located at the top right of the home screen. I also changed the password entry widget from a "Text" widget to an "Entry" widget so that the password can be censored when typed in.